move these to their own model file to avoid circular imports
