+++ posthog/management/commands/run_autoreload_celery.py
@@ -1,5 +1,4 @@
 import django
-
 from django.core.management.base import BaseCommand
 
 from posthog.tasks.utils import CeleryQueue
@@ -14,6 +13,7 @@ class Command(BaseCommand):
     help = "Wrap celery in djangos auto reload functionality"
 
     def handle(self, *args, **options):
+        return
         autoreload.run_with_reloader(self.run_celery_worker)
 
     def run_celery_worker(self):
