+++ posthog/models/__init__.py
@@ -13,11 +13,11 @@ from ..session_recordings.models.session_recording_playlist_item import (
 )
 from ..warehouse.models import DataWarehouseTable
 from ._deprecated_prompts import Prompt, PromptSequence, UserPromptState
-from .alert import Alert
 from .action import Action
 from .action.action_step import ActionStep
 from .activity_logging.activity_log import ActivityLog
 from .activity_logging.notification_viewed import NotificationViewed
+from .alert import Alert
 from .annotation import Annotation
 from .async_deletion import AsyncDeletion, DeletionType
 from .async_migration import AsyncMigration, AsyncMigrationError, MigrationStatus
@@ -47,8 +47,9 @@ from .instance_setting import InstanceSetting
 from .integration import Integration
 from .messaging import MessagingRecord
 from .notebook import Notebook
-from .organization import Organization, OrganizationInvite, OrganizationMembership
+from .organization import Organization, OrganizationMembership
 from .organization_domain import OrganizationDomain
+from .organization_invite import OrganizationInvite
 from .person import Person, PersonDistinctId, PersonOverride, PersonOverrideMapping
 from .personal_api_key import PersonalAPIKey
 from .plugin import (
