+++ posthog/tasks/test/test_email.py
@@ -7,10 +7,15 @@ from freezegun import freeze_time
 
 from posthog.api.authentication import password_reset_token_generator
 from posthog.api.email_verification import email_verification_token_generator
-from posthog.batch_exports.models import BatchExport, BatchExportDestination, BatchExportRun
+from posthog.batch_exports.models import (
+    BatchExport,
+    BatchExportDestination,
+    BatchExportRun,
+)
 from posthog.models import Organization, Team, User
 from posthog.models.instance_setting import set_instance_setting
-from posthog.models.organization import OrganizationInvite, OrganizationMembership
+from posthog.models.organization import OrganizationMembership
+from posthog.models.organization_invite import OrganizationInvite
 from posthog.models.plugin import Plugin, PluginConfig
 from posthog.tasks.email import (
     send_async_migration_complete_email,
